requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
cloudinary==1.38.0
Django==3.2.23
django-cloudinary-storage==0.3.0
djangorestframework==3.14.0
idna==3.6
pillow==10.2.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
